PPAD-23: move helper functions to adopt template
